Move WFSDescribeFeatureHelper from service-wfs to control-base
