+++ control-base/pom.xml
@@ -46,10 +46,6 @@
         </dependency>
         <dependency>
             <groupId>fi.nls.oskari.service</groupId>
-            <artifactId>oskari-wfs</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>fi.nls.oskari.service</groupId>
             <artifactId>oskari-rating</artifactId>
         </dependency>
         <dependency>
